+++ docs/src/examples/QasSearchInput/Basic.vue
@@ -14,20 +14,6 @@ export default {
     return {
       search: ''
     }
-  },
-
-  computed: {
-    entity () {
-      return 'users'
-    },
-
-    fieldsProps () {
-      return {
-        company: {
-          label: 'Alterado label pelo fieldsProps!'
-        }
-      }
-    }
   }
 }
 </script>