+++ ui/src/components/search-box/QasSearchBox.vue
@@ -255,6 +255,7 @@ export default {
 
     setFuse () {
       const list = [...this.list]
+
       this.mx_filteredOptions = list
       this.fuse = new Fuse(list, this.defaultFuseOptions)